+++ platform-sdk/consensus-utility/src/main/java/module-info.java
@@ -6,6 +6,7 @@ module org.hiero.consensus.utility {
     exports org.hiero.consensus.exceptions;
     exports org.hiero.consensus.roster;
     exports org.hiero.consensus.transaction;
+    exports org.hiero.consensus.round;
 
     requires transitive com.hedera.node.hapi;
     requires transitive com.hedera.pbj.runtime;
