+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/consensus/ConsensusRounds.java
@@ -18,6 +18,7 @@ import org.hiero.consensus.model.event.EventConstants;
 import org.hiero.consensus.model.event.NonDeterministicGeneration;
 import org.hiero.consensus.model.hashgraph.ConsensusConstants;
 import org.hiero.consensus.roster.RosterUtils;
+import org.hiero.consensus.round.RoundCalculationUtils;
 
 /**
  * Stores all hashgraph round information in a single place.
